# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Rotas.API",
    "version": "1.0"
  },
  "paths": {
    "/api/Rotas": {
      "get": {
        "tags": [
          "Rotas"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Rotas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdicionarRotaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdicionarRotaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdicionarRotaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Rotas/{id}": {
      "put": {
        "tags": [
          "Rotas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarRotaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarRotaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditarRotaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Rotas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Rotas/calcular-melhor-rota/{origem}/{destino}": {
      "get": {
        "tags": [
          "Rotas"
        ],
        "parameters": [
          {
            "name": "origem",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destino",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdicionarRotaRequest": {
        "required": [
          "destino",
          "origem"
        ],
        "type": "object",
        "properties": {
          "origem": {
            "type": "string",
            "nullable": true
          },
          "destino": {
            "type": "string",
            "nullable": true
          },
          "valor": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "EditarRotaRequest": {
        "required": [
          "destino",
          "origem"
        ],
        "type": "object",
        "properties": {
          "origem": {
            "type": "string",
            "nullable": true
          },
          "destino": {
            "type": "string",
            "nullable": true
          },
          "valor": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  }
}